+++ src/com/jec/commons/jcad/spi/JcadContextImpl.ts
@@ -16,7 +16,7 @@
 
 import {JcadContext} from "../JcadContext";
 import {GuidGenerator} from "../../lang/GuidGenerator";
-import {GuidGeneratorBase} from "../../lang/utils/GuidGeneratorBase";
+import {GlobalGuidGenerator} from "../../lang/utils/GlobalGuidGenerator";
 
 /**
  * The default implementation of the <code>JcadContext</code> interface.
@@ -51,8 +51,7 @@ export class JcadContextImpl implements JcadContext {
    * Initializes this object.
    */
   private initObj():void {
-    let generator:GuidGenerator = new GuidGeneratorBase();
-    this._id = generator.generate();
+    this._id = GlobalGuidGenerator.getInstance().generate();
   }
 
   //////////////////////////////////////////////////////////////////////////////