+++ src/com/jec/commons/jcad/spi/JcadContextManager.ts
@@ -18,7 +18,7 @@ import {JcadContextError} from "../exceptions/JcadContextError";
 import {SingletonError} from "../../exceptions/SingletonError";
 import {Singleton} from "../../lang/Singleton";
 import {GuidGenerator} from "../../lang/GuidGenerator";
-import {GuidGeneratorBase} from "../../lang/utils/GuidGeneratorBase";
+import {GlobalGuidGenerator} from "../../lang/utils/GlobalGuidGenerator";
 import {JcadContext} from "../JcadContext";
 
 /**
@@ -102,8 +102,7 @@ export class JcadContextManager implements Singleton {
    * Initializes this object.
    */
   private initObj():void {
-    let generator:GuidGenerator = new GuidGeneratorBase();
-    this._id = generator.generate();
+    this._id = GlobalGuidGenerator.getInstance().generate();
     this._jcadContextMap = new Map<string, JcadContext>();
   }