+++ src/com/jec/commons/lang/utils/DefaultClassLoader.ts
@@ -15,6 +15,7 @@
 //   limitations under the License.
 
 import {ClassLoader} from "../ClassLoader";
+import {ClassLoaderContext} from "../../context/spi/ClassLoaderContext";
 
 /**
  * The <code>DefaultClassLoader</code> class is the default implementation of
@@ -39,7 +40,9 @@ export class DefaultClassLoader implements ClassLoader {
    * @inheritDoc
    */
   public loadClass(path:string):any {
+    ClassLoaderContext.getInstance().setPath(path);
     let classRef:any = require(path);
+    ClassLoaderContext.getInstance().setPath(null);
     let props:string[] = Object.keys(classRef);
     let className:string = props[0];
     return classRef[className];